Updates the SQL class and adds update_Daily method to main class
This update changes the WxChall_SQL class so that it is in
charge of generating the file path to the database.

Also adds the update_Daily method to the WxChallenge class,
which is designed to be called, you guessed it, daily to update
the forecast results.

The _date attribute of WxChallenge has also been changed to
be one day sooner than todays date. This is because the forecast
that verifies for the Monday forecast (i.e., forecast for Tuesday)
is marked as day 1 of the challenge, i.e., the start date of the
city. This may have to be changed when the game actually starts and
I get an better idea of when the forecasts are released.
  Perhaps a better way would be to check to see how many days for
a given city exist in the SQL table and then go from there.

v0.0.2 05 Sep. 2018 1534 CDT