Adds forecaster class and cleans up some code
This version adds a 'forecaster' class, wherein a forecaster
is defined using name, category, school, semester, and year.
There is a method that allows for adding forecasts IF the
five (5) attributes of the class mentioned above match that
of the forecast. A call to the 'get_forecasts' method returns
a list of 'forecaster' instances. This makes it eaiser to
parse the data.

In each forecaster instance, there are attributes that match
the columns of the SQL data, with lists of the data that are
sorted based on date.

Also updates the utils.scoring function to handle the updated
list of forecasters/forecasts.

v0.0.3 20180907 0840 CDT