Move to pandas.DataFrame and other updates
The main change in this update is the utilization of pandas
DataFrames. DataFrames are used (almost) everywhere data is
begin handled. When forecasts are read from th SQL database,
they are converted to a DataFrame. All the unique forecasters
in the DataFrame are then identified, and forecaster objects
(now subclasses of the DataFrame class) are initialized.

A calc_grade method has been added to the forecaster class.
This method loops over all the unique station identifiers in
the DataFrame and determines: # forecasts missed, # forecasts
worse than cimatology (if input), and if beat local and/or
national consensus (if input). These data are then stored in
a DataFrame under the 'grades' attribute.

Not sure (at this point) what other functionality will be useful.
More will change when game actually starts.

v0.1.0 20180907 1651 CDT