Adds daily updater command line util
This version adds the WxChall_Daily_Update command that is to
be run, daily, in a cronjob type style to update the forecast
database.

Version also updates the update_Daily method of the WxChallenge
class to check the date three (3) days before todays date for
the verifications; it takes at least 1 day for the verification
to appear so this gives some extra time.

v0.1.2 24 Sep. 2018 1620 CDT