Adds new updating methods and improves SQL database updating
This update adds teh update_Day method to the WxChallenge class.
This method allows one to update data for a specific day of the
challenge using semester, year, station ID, and forecast day.

Improvements were made to forecast updating/inserting in the
SQL table. Before, a check was done to see if every part
of a forecast match an exising forecast. If there were no
matches found, then a new forecast was entered. This could
cause mulitple 'verifications' of a forecast as a day
progresses and ranks, errors, etc. change. Thus, a check
is now done using forecaster ID, school, category, and
date to see if a forecast exists. If one exists, then
attributes of the forecast that differ are updated in
the SQL table; if all attributes are the same then
nothing is changes. If the forecast is NOT found, then a
new forecast is added to the SQL table.

Also updates sorting of forecaster list returned by
get_forecasts method. Was sorting with upper case
first, then all lower case; updated to case insensitive.

v0.1.3 28 Sep. 2018 1558 CDT