Fixes issue in the WxForecast.calc_grades method
The issue arose from computing a sum of all the missed forecast.
This would cause someone who missed to first forecast of the period
to 'miss' all forecasts or any combination of events. The fix made
was to take the maximum value of the 'abs' column instead of the sum.

v0.2.3 20181022 1716 CDT