Fixes schedule issues on update and improves code flow
This update (should) fix the issue with the schedule not updating
throughout the semester as the current date is greater than the
previously available forecast cities. If the end date of the latest
city is less than the current date, then update the schedule.

Also updates some of the backend data. URL information is moved
to its own class, WxChall_URLs, which will make it easier to update
URL paths if they change in the future. The checkURL util function
has also been added as a method to the WxChall_URLs class

A WxChall_Schedule class has also been created to take care of
schedule information. The schedule information is still loaded/
downloaded in the WxChall_SQLite class, but it (can be) parsed
in the WxChall_Schedule and update using methods in that class.

v0.2.4 2018 Oct 23 1546 CDT