Updates calc_grades method of forecasts
This update changes the calc_grades method of the forecasts
class. The method now gives a zero (0) score for missing
all forecast times in a given city. There are 2 free days
given by default (i.e., you only have to forecast 6 out of
the 8 days to get a 100). If you do NOT beat climatology for
the 6 days you forecast, your score will drop to 70.

Bonus points are awarded based on your national normalized
cumulative error for each city AND the school normalized
cumulated error for each city.

The private __addFinalScore method of the ExcelBook class
has also been updated so that the total bonus national and
school bonus points are on the 'Final Grades' sheet.

v0.2.7 20181211 2204 CDT