Fixes spring roster issue; more logging
Because the roster for the spring semester includes forecasters
who only registered for the Fall semester, the code would break
as there are no forecasts in the spring semester for students
that only forecasted in the fall. To fix this issue, the semester
name (i.e., spring or fall) is passed into the fix_Roster_CSV
function. If the semester is in the spring, then we remove any
forecaster from the roster that was registered for only fall;
there will be an 'F' in the fall_spring_both column.

Logging is improved slightly by initializing the package logger
in the main __init__.py file and adding logging to the classes
in WxChall_Excel.py

v0.2.9