Cleans up code in CLIs
Argument adding to the argparser in the CLIs is formatted to be
more readable; each keyword given own line.

Added commented out code in setup.py that will enable moving
of .sql file to home directory on update and then moving file
back to correct location on install. This is likely not needed
though because .sql file is not shipped with package. and is
therefore not touch on update

v0.2.16