Adds XLSX file for all forecasters
In previous version, XLSX files were only created for classes.
This means that if a forecaster was not part of a class, their
'grade' was never computed. A file is now generated in addition
to the class files that contains all forecasters, whether they
participated as part of class or not.

Also fixes bug in the WxChall_Grades CLI that was introduced when
I added the ability to select different verbosity levels.

v0.2.18