Small bug fixes for date parsing and schedule
There were some issues with schedule downloading logic where if
the code is not run for every semester one could have gaps in the
schedule. Have added the --full flag the the Semester Update CLI.

Issues with some date parsing in verification where the website
shows dates for 31 of February, which is wrong. Changed to using
first date in the verification list and building to given date
based on number of rows in table. Hopefully this doesn't break
later.

Adds missing 'requests' dependency to the setup file

v0.3.3